python 3month internship
This project is a fun and easy-to-play Memory Puzzle game built with Python’s pygame library. The aim is to match pairs of hidden cards by remembering where they are. When you click on a card, it flips over to show a picture or pattern. You keep flipping two cards at a time, trying to find matching pairs. It’s a great way to challenge your memory and focus, making it enjoyable for players of all ages. Plus, it’s a nice project for anyone looking to improve their Python skills with simple game development.